ヘッダーをコンポーネントに
