+++ client/components/Header/PortalHeader.module.css
@@ -1,0 +1,83 @@
+.header {
+  position: sticky;
+  top: 0;
+  z-index: 1000;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0.75rem;
+  color: white;
+  background-color: #1e40af;
+}
+
+.mainTitle {
+  font-size: 1.25rem;
+  font-weight: 700;
+}
+
+.menuButton {
+  padding: 0.5rem;
+  font-size: 1.5rem;
+  color: white;
+  cursor: pointer;
+  background: none;
+  border: none;
+}
+
+.menuIcon {
+  width: 1.5rem;
+  height: 1.5rem;
+}
+
+.nav {
+  position: absolute;
+  top: 100%;
+  right: 0;
+  left: 0;
+  z-index: 999;
+  display: none;
+  flex-direction: column;
+  background-color: #1e3a8a;
+}
+
+.navOpen {
+  display: flex;
+}
+
+.navLink {
+  padding: 1rem;
+  font-size: 1rem;
+  color: white;
+  text-decoration: none;
+  border-bottom: 1px solid rgb(255 255 255 / 10%);
+}
+
+.navLink:last-child {
+  border-bottom: none;
+}
+
+@media (width >= 640px) {
+  .header {
+    padding: 1rem;
+  }
+
+  .mainTitle {
+    font-size: 1.5rem;
+  }
+
+  .menuButton {
+    display: none;
+  }
+
+  .nav {
+    position: static;
+    display: flex;
+    flex-direction: row;
+    background-color: transparent;
+  }
+
+  .navLink {
+    padding: 0.5rem 1rem;
+    border-bottom: none;
+  }
+}
